refactor: remove hook and fix a couple of typing errors (#13425)
* refactor: rename hook and fix a couple of typing errors

* refactor: remove custom hook, replace with useRef, fix typing errors

* chore: fix some stuff

* change variable in AudioAsset

* chore: add captions to audio

* chore: rename variable

* chore: code review tasks

* fix: use useSate for viewportObservers and resizeObservers instead of useRef

+++ src/script/conversation/ConversationRepository.ts
@@ -127,6 +127,7 @@ import {extractClientDiff} from './ClientMismatchUtil';
 import {Core} from '../service/CoreSingleton';
 import {ClientState} from '../client/ClientState';
 import {MLSReturnType} from '@wireapp/core/src/main/conversation';
+import {isMemberMessage} from '../guards/Message';
 
 type ConversationDBChange = {obj: EventRecord; oldObj: EventRecord};
 type FetchPromise = {rejectFn: (error: ConversationError) => void; resolveFn: (conversation: Conversation) => void};
@@ -619,24 +620,29 @@ export class ConversationRepository {
     const mappedMessageEntities = await this.addEventsToConversation(events, conversationEntity);
     conversationEntity.hasAdditionalMessages(hasAdditionalMessages);
     if (!hasAdditionalMessages) {
-      const firstMessage = conversationEntity.getFirstMessage() as MemberMessage;
-      const checkCreationMessage = firstMessage?.isMember() && firstMessage.isCreation();
-      if (checkCreationMessage) {
-        const groupCreationMessageIn1to1 = conversationEntity.is1to1() && firstMessage.isGroupCreation();
-        const one2oneConnectionMessageInGroup = conversationEntity.isGroup() && firstMessage.isConnection();
-        const wrongMessageTypeForConversation = groupCreationMessageIn1to1 || one2oneConnectionMessageInGroup;
+      const allMessages = conversationEntity.getAllMessages();
+      if (allMessages.length) {
+        const firstMessage = allMessages[0];
+        if (isMemberMessage(firstMessage)) {
+          const checkCreationMessage = firstMessage.isMember() && firstMessage.isCreation();
+          if (checkCreationMessage) {
+            const groupCreationMessageIn1to1 = conversationEntity.is1to1() && firstMessage.isGroupCreation();
 
-        if (wrongMessageTypeForConversation) {
-          this.messageRepository.deleteMessage(conversationEntity, firstMessage);
-          conversationEntity.hasCreationMessage = false;
-        } else {
-          conversationEntity.hasCreationMessage = true;
+            const one2oneConnectionMessageInGroup = conversationEntity.isGroup() && firstMessage.isConnection();
+            const wrongMessageTypeForConversation = groupCreationMessageIn1to1 || one2oneConnectionMessageInGroup;
+
+            if (wrongMessageTypeForConversation) {
+              this.messageRepository.deleteMessage(conversationEntity, firstMessage);
+              conversationEntity.hasCreationMessage = false;
+            } else {
+              conversationEntity.hasCreationMessage = true;
+            }
+          }
+
+          if (!conversationEntity.hasCreationMessage) {
+            this.addCreationMessage(conversationEntity, this.userState.self().isTemporaryGuest());
+          }
         }
-      }
-
-      const addCreationMessage = !conversationEntity.hasCreationMessage;
-      if (addCreationMessage) {
-        this.addCreationMessage(conversationEntity, this.userState.self().isTemporaryGuest());
       }
     }
     return mappedMessageEntities;
